+++ apps/jaspr_pad/lib/providers/dart_service_provider.dart
@@ -1,8 +1,9 @@
+import 'package:dart_mappable/dart_mappable.dart';
 import 'package:http/http.dart' as http;
 import 'package:jaspr_riverpod/jaspr_riverpod.dart';
 
 import '../adapters/html.dart';
-import '../main.container.dart';
+import '../main.init.dart';
 import '../models/api_models.dart';
 import '../models/sample.dart';
 import '../models/tutorial.dart';
@@ -10,7 +11,9 @@ import '../models/tutorial.dart';
 final dartServiceProvider = Provider((ref) => DartService(ref));
 
 class DartService {
-  DartService(this.ref);
+  DartService(this.ref) {
+    initializeMappers();
+  }
 
   final Ref ref;
   final client = http.Client();
@@ -29,12 +32,14 @@ class DartService {
 
   Future<T> _get<T>(String path) async {
     var response = await client.get(Uri.parse('${window.location.origin}/api/$path'));
-    return mainContainer.fromJson<T>(response.body);
+    return MapperContainer.globals.fromJson<T>(response.body);
   }
 
   Future<T> _request<T>(String action, Object body) async {
-    var response =
-        await client.post(Uri.parse('${window.location.origin}/api/$action'), body: mainContainer.toJson(body));
-    return mainContainer.fromJson<T>(response.body);
+    var response = await client.post(
+      Uri.parse('${window.location.origin}/api/$action'),
+      body: MapperContainer.globals.toJson(body),
+    );
+    return MapperContainer.globals.fromJson<T>(response.body);
   }
 }