+++ examples/riverpod_app/lib/main.dart
@@ -4,5 +4,5 @@ import 'app.dart';
 
 void main() async {
   // runs the app component on both server and client
-  runApp(Document.app(body: App()));
+  runApp(Document(body: App()));
 }