+++ experiments/platform_test/web/main.clients.dart
@@ -1,0 +1,14 @@
+// GENERATED FILE, DO NOT MODIFY
+// Generated with jaspr_builder
+
+import 'package:jaspr/browser.dart';
+import 'components/app.client.dart' deferred as i0;
+
+void main() {
+  registerClients({
+    'components/app': loadClient(
+      i0.loadLibrary,
+      (p) => i0.getComponentForParams(p),
+    ),
+  });
+}