+++ packages/jaspr/lib/src/foundation/annotations.dart
@@ -1,18 +1,11 @@
 import '../framework/framework.dart';
 
-class AppAnnotation {
-  const AppAnnotation._();
+class ClientAnnotation {
+  const ClientAnnotation._();
 }
 
-/// Used to annotate an app component
-const app = AppAnnotation._();
-
-class IslandAnnotation {
-  const IslandAnnotation._();
-}
-
-/// Used to annotate an island component
-const island = IslandAnnotation._();
+/// Used to annotate a client component
+const client = ClientAnnotation._();
 
 enum ImportPlatform { web, server }
 
@@ -60,24 +53,7 @@ abstract class ComponentEntryMixin<T extends Component> {
 
 class ComponentEntry<T extends Component> {
   final String name;
-  final ComponentEntryType type;
   final Map<String, dynamic>? params;
 
-  ComponentEntry.app(this.name, {this.params}) : type = ComponentEntryType.app;
-  ComponentEntry.island(this.name, {this.params}) : type = ComponentEntryType.island;
-  ComponentEntry.appAndIsland(this.name, {this.params}) : type = ComponentEntryType.appAndIsland;
-
-  bool get isApp => type == ComponentEntryType.app || type == ComponentEntryType.appAndIsland;
-  bool get isIsland => type == ComponentEntryType.island || type == ComponentEntryType.appAndIsland;
-}
-
-enum ComponentEntryType {
-  app(true, false),
-  island(false, true),
-  appAndIsland(true, true);
-
-  final bool isApp;
-  final bool isIsland;
-
-  const ComponentEntryType(this.isApp, this.isIsland);
+  ComponentEntry.client(this.name, {this.params});
 }