+++ packages/jaspr_builder/lib/src/utils.dart
@@ -3,16 +3,14 @@
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/element.dart';
-import 'package:jaspr/src/foundation/annotations.dart' show AppAnnotation, Import, IslandAnnotation;
+import 'package:jaspr/src/foundation/annotations.dart' show ClientAnnotation, Import;
 import 'package:jaspr/src/framework/framework.dart' show Component, Key;
-
 import 'package:source_gen/source_gen.dart';
 
 const String generationHeader = "// GENERATED FILE, DO NOT MODIFY\n"
     "// Generated with jaspr_builder\n";
 
-var appChecker = TypeChecker.fromRuntime(AppAnnotation);
-var islandChecker = TypeChecker.fromRuntime(IslandAnnotation);
+var clientChecker = TypeChecker.fromRuntime(ClientAnnotation);
 var componentChecker = TypeChecker.fromRuntime(Component);
 final keyChecker = TypeChecker.fromRuntime(Key);
 var importChecker = TypeChecker.fromRuntime(Import);