+++ packages/jaspr_cli/tool/templates/islands/__brick__/web/components/content.island.dart
@@ -1,9 +1,0 @@
-// GENERATED FILE, DO NOT MODIFY
-// Generated with jaspr_builder
-
-import 'package:jaspr/browser.dart';
-import 'package:{{name}}/components/content.dart' as a;
-
-Component getComponentForParams(ConfigParams p) {
-  return a.Content(greet: p.get('greet'));
-}