+++ packages/jaspr_router/lib/src/misc/errors.dart
@@ -1,0 +1,11 @@
+/// Thrown when [Router] is used incorrectly.
+class RouterError extends Error {
+  /// Constructs a [RouterError]
+  RouterError(this.message);
+
+  /// The error message.
+  final String message;
+
+  @override
+  String toString() => 'RouterError: $message';
+}